Refactoring serialization of kvs_payloads
Originally we serialize the vector of kvs_payloads as a protobuf object.
Now, we convert the vector to a delimeted string and serialize it as a
string field in the CapsuleBlock protobuf. This was done in effort to
save memory, but has yielded no improvements in the test. Currently, the
test still segfaults at level2 (about 1000 pairs)